separated entity search manage menu from search action bar"
