+++ ui/src/components/common/SearchActionBar.scss
